New logic for hardcore mode
Clears the users score if they lose 3 in a row. 
This is managed with resetScoreHC in the main file which checks if we are hardcore when we get 3x wrong guesses in a row, and resets the score if we are in HC mode. 
In utils.asm, the 'newGame' section of the increment_attempts procedure resets the score printed on screen to 0 when user loses in hardcore mode.